Refactor UI and improve VPN logic
